optimized structure three folder
